finished spring security configuration with jwt

+++ src/main/java/com/example/demo/repo/UserRepo.java
@@ -6,8 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.domain.User;
 
-public interface UserRepo extends JpaRepository<User, Integer> {
-
+public interface UserRepo extends JpaRepository<User, Long> {
   Optional<User> findByUsername(String username);
-
 }
